requirements.txt
flask
requests
gunicorn
flask-limiter
redis>=3.0
dnspython